Properly implemented the [Delete] key
I recently found out that the delete key deletes the character in
front of the cursor instead of working like backspace, so I
went back to implement it here.

- Added static function remove_from_string_2
- Modified prompts to call remove_from_string_2 when KEY_DC

Probably doesn't work with Mac because of the different keyboard,
need to get one of my Mac using friends to test and help me out.